+++ framework/Config/Config.php
@@ -1,0 +1,28 @@
+<?php
+
+namespace Framework\Config;
+
+class Config
+{
+    private static ?array $env = null;
+
+    public static function env(string $name, mixed $default = null): mixed
+    {
+        self::readEnv();
+
+        if (!array_key_exists($name, self::$env)) {
+            return $default;
+        }
+
+        return self::$env[$name];
+    } 
+
+    /** Read env file */
+    private static function readEnv(): void
+    {
+        if (self::$env !== null) {
+            return;
+        }
+        self::$env = ConfigReader::readFile(__DIR__ . '/../../.env');
+    }
+}