+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Types of changes: `Added`, `Changed`, `Deprecate`, `Removed`, `Fixed`, `Secruity
 - Added command to create a new controller
 - Added all CRUD actions for product with ProductController
 - Added created and updated timestamps to migrations and users table
+- Added Config class to make access to env settings easier
 
 ### Changed
 - Improved the autoloader to work for the new folder structure of version 2